requirements.txt
# Data Manipulation and Analysis
pandas==2.1.4
numpy==1.26.2
scipy==1.11.4

# Visualization
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.18.0
kaleido==0.2.1

# Machine Learning - Traditional
scikit-learn==1.3.2
xgboost==2.0.3

# Time Series Models
statsmodels==0.14.1
prophet==1.1.5

# Deep Learning
tensorflow==2.15.0

# Web Framework
streamlit==1.29.0

# Data Collection
requests==2.31.0
beautifulsoup4==4.12.2

# Utilities
openpyxl==3.1.2
joblib==1.3.2
python-dateutil==2.8.2
pytz==2023.3

# Geospatial (for India map visualizations)
geopandas==0.14.1
folium==0.15.1

# Progress bars
tqdm==4.66.1

# Testing
pytest==7.4.3
